requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
faiss-cpu==1.7.3
numpy==1.24.3
together==1.2.1
python-dotenv==1.0.0
transformers==4.30.2
boto3==1.26.0